+++ Flow/Changes.md
@@ -13,6 +13,8 @@
 ### New methods
 - **Lockable**
   - Added `def unlocked = !locked`
+- **Seq** (**CollectionExtensions**)
+  - Added `.headPair`, `.lastPair`, `.uniquePairsIterator` and `.pairFrom(Int)` utility functions
 - **ThreadPool**
   - Added `.stop()`, which may be used for clearing all threads (including the core threads) as soon as they're freed
 ### Other changes